+++ modules/admin/tests/data/fixtures/TagFixture.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace admintests\data\fixtures;
+
+use yii\test\ActiveFixture;
+
+class TagFixture extends ActiveFixture
+{
+    public $modelClass = 'luya\admin\models\Tag';
+
+    public function getData()
+    {
+        return [
+            'tag1' => [
+                'id' => '1',
+                'name' => 'john',
+            ],
+            'tag2' => [
+                'id' => '2',
+                'name' => 'Jane',
+            ],
+        ];
+    }
+}